pyproject.toml
[project]
name = "AlphaZero_in_Pan"
version = "0.1.0"
description = "Using AlphaZero algorithm to play the game of Pan"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.3",
    "orbax>=0.1.9",
    "pydantic>=2.12.5",
    "pygame>=2.6.1",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.1",
    "wandb>=0.22.2",
]


[dependency-groups]
ml = [
    "flax>=0.12.0",
    "jax>=0.7.2",
    "optax>=0.2.6",
]
dev = [
    "pre-commit>=4.5.0",
    "pyright>=1.1.407",
    "ruff>=0.6.4",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

# Ruff configuration
[tool.ruff]
target-version = "py312"
line-length = 150

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # import sorting
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "lf"
